shun-speedtest-graph/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
apscheduler==3.10.4
speedtest-cli==2.1.3
tweepy==4.14.0
